Enhance tests & modernize code for collections/utilities
- Add extensive unit tests for ConcurrentHashSet<T> (edge cases, interface compliance, concurrency, error handling)
- Refactor and expand PerformanceStopwatch and related tests
- Use discard assignments for unused return values in tests
- Use C# 12 array literals and concise conditionals in core extensions
- Improve test resource management in PathHelperTests
- Update file headers with new modification dates
- General code cleanup and improved readability throughout

+++ source/Unit Tests/dotNetTips.Spargine.Extensions.Tests/EnumerableExtensionsTests.cs
@@ -1339,7 +1339,7 @@ public class EnumerableExtensionsTests
 
 		var result = numbers.FastLongCount(n => n % 2 == 0);
 
-		Assert.AreEqual((long)(Count / 2), result);
+		Assert.AreEqual((Count / 2), result);
 	}
 
 	[TestMethod]
